Show a delete button if multiple volumes are selected
Fixes issue OP-884. Shows a delete button instead of an add button.

Although fixes the issue I came across of deleting nothing (no selection).
If the array of multiple selections has already been initialized, it will
be an array and not null as in the beginning if you deselect all volumes.

VolumesCtrl.js:

* Prevents false positive multiple selections.

volumes.html:

* Added delete button if multiple volumes are selected.


Signed-off-by: Stephan Müller <stephan@open-attic.org>