oaconfig-install.sh: Don't enable the NPCD broker module for Nagios 4 (OP-820)
The EPEL package repository for RHEL7/CentOS7 recently updated Nagios to
version 4, which broke loading the pnp4nagios module.

Now oaconfig-install.sh checks the Nagios binary's version number and skips
enabling npcdmod.o when version 4 is detected. Icinga 1.13 as used by SUSE and
Nagios 3 as used by Debian/Ubuntu continue to work as before.

This fixes the issue that "oaconfig install" would hangs with the message
"Waiting for status.dat to appear...". As a next step, the Nagios 4
configuration needs to be changed to enable using pnp4nagios in bulk mode
with npcd.

Signed-off-by: Lenz Grimmer <lenz@openattic.org>