Updated debian/default/openattic to include missing config variables (OP-820)
The previous commit that made oaconfig-install check for the Nagios version
made use of the NAGIOS_BINARY_NAME variable, which did not exist in the
Debian default config file. This resulted in an error when running "oaconfig
install" ("NAGIOS_BINARY_NAME: unbound variable"). Added missing config options
from the openattic default config files used by the RPM packages and adjusted
them to match the Debian environment. In a next step, the redundant settings
should be removed (which is tracked in OP-944).

Signed-off-by: Lenz Grimmer <lenz@openattic.org>