oaconfig: removed support for "oaconfig install <package>" (OP-659)
"oaconfig install <package>" was hard-coded to run "apt-get" for installing
packages, which only worked on Debian/Ubuntu and caused errors on RPM-based
distributions like RHEL/CentOS or SUSE.

Instead of adding cross-distro support for all the various package management
utilities, we try to follow KISS and simply drop this feature and refer to the
respective package manager. It should not be the task of "oaconfig" to perform
package installations.

Signed-off-by: Lenz Grimmer <lenz@openattic.org>