Installation: Fixed check_openattic_rpcd Nagios plugin (OP-977)
The Nagios plugin "check_openattic_rpcd" tried to call another Nagios plugin
"check_tcp" using a hard-coded path name that did not exist on EL7.
Added a configuration setting NAGIOS_PLUGIN_DIR to the default config file and
changed to plugin to use that setting instead of a fixed path.
Added "nagios-plugins-tcp" as an additional requirement to the
openattic-module-nagios RPM package, to ensure the plugin is actually installed.

Signed-off-by: Lenz Grimmer <lenz@openattic.org>