Installation: Fixed Nagios plugin Python path issues on EL7 (OP-820)
Unfortunately the "nagios" user on EL7 does not seem to inherit the
PYTHONPATH environment variable defined in /etc/profile.d/openattic.sh and
therefore the Python-based Nagios plugins fail to read the distro settings.

As a stop-gap measure, we now set the path to the openATTIC directory explicitly.
Ideally, the path should be obtained from the OADIR config option in
/etc/(default|sysconfig)/openattic

Signed-off-by: Lenz Grimmer <lenz@openattic.org>