Installation: Added tmpfiles.d configuration file (OP-738)
On some systems (e.g. EL7, SLES12), /var/lock is a symbolic link
to /run/lock, which is mounted on tmpfs and thus disappears after
a reboot. This confuses openATTIC (no disks, pools or volumes
are visible after a reboot) and must be avoided. The recommended
way is to instruct systemd to create the directory at bootup
time using the tmpfiles.d mechanism.

Signed-off-by: Lenz Grimmer <lenz@openattic.org>