Extend the pagination with an input field
With the input field you can change the page by typing an number.
It will instantly load the page as you type.

Removed paginatorService.js because it wasn't needed anymore.
Add a new variable to the pagination scope to hold the number that is displayed in the input field.

Changed the switchPage function to set displayNumber and page accordingly.

Changed the pagination template to have an input field with the current page number (human readable (page + 1)).

Added new css to center the input field.



Signed-off-by: Stephan Müller <stephan@open-attic.org>