Fixed dependency on the Python RADOS module for the ceph module DEB (OP-1028)
While the upstream Ceph project provides the RADOS Python bindings in a separate
Debian package "python-rados", stock Debian and Ubuntu ship these bindings as
part of the "python-ceph" package. To support installing openattic-module-ceph
on both Debian and Ubuntu by using either the distribution's Ceph packages or
the upstream apt repository, having a dependency on "python-ceph" is the safest
bet.

Signed-off-by: Lenz Grimmer <lenz@openattic.org>