Backend: Enable detecting more than 16 SCSI disks (OP-1136)
The "update_disks()" method is called during "oaconfig install", to detect all
hard disks attached to a system. It checked if a device's block major number was
8, which made it only detect the first 16 SCSI devices on a system.
The code now checks for all possible SCSI disk major device numbers, by
comparing the major number against a list of possible numbers.
Improved the loop that scans all devices by filtering it for block devices of
type disk only, removed now obsolete checks (e.g. using the device's minor
number to determine between the actual disk and its partitions, or if a device's
ID_TYPE matches "disk" (which it does if it's major number is in the list).

Signed-off-by: Lenz Grimmer <lenz@openattic.org>