Separate the revision argument to a channel and a revision
The channel argument is mandatory. Either "stable" or "unstable" has to be used as channel. The revision is a separate argument by now, which is optional. If it's ommited, "stable" results in the latest labeled tag on the default branch being used. In case of "unstable", the tip of the default branch is used. New is, that a revision may be provid additionally and not instead of the release channel. This results in being able to create a stable or unstable tarball out of every given mercurial tag (no tag revisions aren't supported so far).

Signed-off-by: Patrick Nawracay <patrick.nawracay@it-novum.com>