Documentation: Improved installation instructions
Added a note to the storage configuration section that this step can be
skipped if one just wants to use openATTIC for managing a Ceph cluster.

Moved the step of importing the GPG build key before adding the apt repos,
to avoid an error about a missing key if one runs `apt-get update` right
after adding the repo.

Signed-off-by: Lenz Grimmer <lenz@openattic.org>