Packaging: added new `taskqueue` backend code to `openattic-base` DEB (OP-1414)
The newly added `taskqueue` backend module was not listed in the
`openattic-base.install` file, so it was missing from the `openattic-base` DEB
package. This caused an update or installation using the nightly builds to fail
with `ImportError: No module named taskqueue.models`. Added the files belonging
to this module to the `openattic-base` DEB package to fix this.

Signed-off-by: Lenz Grimmer <lenz@openattic.org>