Installation: openattic.log must be writable for the openattic group (OP-1320)
The openATTIC log file /var/log/openattic/openattic.log must be owned
by openattic:openattic and must be group-writable, in order to allow
the Nagios plugin (which is run under the "nagios" user that also belongs
to the "openattic" user group) to log messages into this file. Fixed the
ownership and permissions for the RPM and DEB packages and also updated
the logrotate configuration to maintain proper ownerships and permissions.

Signed-off-by: Lenz Grimmer <lenz@openattic.org>