OP-1549: Set nodb context before calling NodbModel.get_all_objects.
Reason: When generating the HTML REST API form, django.forms.models.ModelChoiceIterator
accesses field.queryset, which is kind of static and thus has no context yet (and must never have one!).

This is not a really nice solution, as this now depends the global variable NodbManager.nodb_context,
which can be set by everyone. But I don't see a better way to deal with it at this point.

Signed-off-by: Sebastian Wagner <sebastian.wagner@it-novum.com>