Backed out changeset 11a914a9ee3d
Since we package the nagios plugins to the nagios dir. itself in the rpm spec.
this changeset wouldn't work on its own unless we move back the icinga plugins
to its own dir. However since icinga is compatible with nagios and looks for
plugins there as well, and the only place NAGIOS_PLUGIN_DIR seemed to be used
for was for host checking, the previous change went undetected.

Signed-Off-By: Abhishek Lekshmanan <abhishek@suse.com>