Extend `make_dist.py` to be able to set an hg tag (OP-1374)
This is especially useful for a release. The tag can be created alongside the tarball and the changes may be pushed back to the repository. This enables the script to be used to automate the release process as it takes care of steps which otherwise had to be done manually.

Signed-off-by: Patrick Nawracay <patrick@openattic.org>