Installation: fixed location of the Nagios `status.dat` file on EL7 (OP-1955)
Nagios was updated from version 4.0.8 to 4.2.4 in EPEL7, which also introduced
a number of changes to the file system layout. In particular, the `status.dat`
file was moved from `/var/log/nagios/` to `/var/spool/nagios`, which causes the
installation via `oaconfig install` to break. Updated the path in the EL7 config
file accordingly.

Signed-off-by: Lenz Grimmer <lenz@openattic.org>