Installation: fixed "OAUSER: unbound variable" error during update (OP-2046)
The configuration file /etc/default/openattic was updated for 2.0.19. By default
RPM did not replace the file, if the user modified it manually. Instead, RPM
will now replace it with the upstream version, keeping a backup copy of the
user-modified file around. This ensures that new settings required by oaconfig
are actually in place when oaconfig is run after updating to 2.0.19.
Added a few safeguards to oaconfig to make it fail more gracefully in case the
setting does not exist in the config file for some reason.

Signed-off-by: Lenz Grimmer <lenz@openattic.org>