Implement a RGW Admin Ops API proxy layer (OP-1612)
This RGW proxy takes care of the S3 authentication and offers the web
frontend a resource which may be used to talk to the admin resource of
the configured RGW host.

The configuration needs to be made in the `backend/settings.py`. The
variable names are:

 - RGW_HOST
 - RGW_PORT
 - RGW_ACCESS_KEY
 - RGW_SECRET_KEY

If a configuration value is missing, the API will return a HTTP status
code 500 and a JSON body. The returned JSON object will contain a `Code`
field, equally to the RGW Admin Ops API in case of an error. The code
will be `ConfigurationIncomplete`.

The admin resource on the Rados Gateway is configurable using the
RGW_ADMIN_RESOURCE configuration variable. It defaults to "admin".

Note that the web UI has to use it's usual authentication method
(including CSRF for POST/PUT/DELETE requests).

The API resource will be available using the `api/rgw` resource.

Resolves: OP-1612
See also: OP-2153, OP-2154

Signed-off-by: Patrick Nawracay <pnawracay@suse.com>