Grafana integration (OP-829)
We used to use nagios to display statistic graphs of ceph pools. I replaced those nagios based graphs with a grafana integration. Furthermore, I added grafana to the openATTIC dashboard view, which can be used by selecting Grafana in the dropdown. If there is no Grafana dashboard configured, just add it with exactly this name. OSDs and Nodes can be monitored as well.

Signed-off-by: Sebastian Krah <skrah@suse.com>