Select and check the feature case on RBD creation
With this changes the RBD creation will work as expected. Before this
change the creation only tested for enabled features after the creation,
that mend that the test has to select all features before the creation
call and that it wasn't checked if only the feature case was enabled.
Now RBD creation will select the feature case for you and it'll check
if features that were not enabled are disabled after the creation.

Signed-off-by: Stephan Müller <smueller@suse.com>