Rewrite the RBD creation suite
Rewrote the RBD creation suite, because it only did the same over and
over again only stopping the test to be fast. Now the test won't test
the creation of every RBD feature combination, but it will combine the
object size testing and the RBD feature case testing and only create
three RBDs and not 67 RBDs. Of course this also mean that we have 64*2
less test cases but the three chosen ones are enough. The first tests
one feature and the smallest size, the second the average size and the
default case and the last will create a RBD with the greatest size and
all options activated that can be activated.

Signed-off-by: Stephan Müller <smueller@suse.com>