Add a new widget for MONs (OP-1933)
Created a new widget for MONs. The MONs are displayed in a similar way
as in the OSD widget. The difference is that each MON is displayed
instead of all MONs in a specific state. On hover each MON displays it's
name, state and IP address in a popover. If the MON is an unhealthy state
it'll also show details about the failure.

Signed-off-by: Stephan Müller <smueller@suse.com>