Packaging: Consolidate subpackages into a single `openattic` RPM (OP-2342)
The original packaging was way too fine-grained and made the packaging
and installation unnecessarily complex. As a lot of the subpackages had
to be installed in any case, the separation made little sense. The focus
on Ceph management in openATTIC 3.x resulted in even fewer optional
packages.

Improve the RPM spec file to be much simpler and easier to maintain.

Merge the following subpackages and the scripts into the main
`openattic` RPM:

* openattic-base
* openattic-gui
* openattic-module-ceph
* openattic-module-ceph-deployment
* openattic-module-icinga
* openattic-pgsql

Clean up the file list and list of dependencies, update the package
description.

Signed-off-by: Lenz Grimmer <lenz@openattic.org>