Resets graph data if needed (OP-2358)
The reason of OP-2358 is that if no data exists the API throws an
error instead of giving back an empty object.

Now the graph factory is called if the call succeeded or failed. The
graph factory now resets the graph if no data is delivered.

Preventing the error message will also solve OP-2002.

Signed-off-by: Stephan Müller <smueller@suse.com>