Close displayed unsaved changes dialogs (OP-2189)
Tests were failing a lot if a item that the tests wanted to create
already existed, because the e2e suite got stuck on the unsaved
changes dialog. I added a page object for the dialog that is used by
a common function with a self explaining name. Also added two new
functions that will help to set a new location or just leave the form
through the back button.

Signed-off-by: Stephan Müller <smueller@suse.com>