RBD-form: EC overwrite pools, crush rules for all and pg calculation
This commit was squashed because of integrity reasons. This commit will
fix multiple issues on the same file.

Commit 1:
Enables creating EC overwrite pools in the UI (OP-2357)

Created the first erasure coded pool flag "overwrite" in the pool form
which will only been shown if there are bluestore OSDs available. Added
tests to verify it.

Commit 2:
Enables crush rule set for all pool types (OP-2378)

If only one cluster exists the cluster selection will be hidden and the
cluster will be selected. After selecting a type the form will validate
if crush rules for the selected type are existing, if not the selection
will not be shown. If only one exist the one will be selected and the
selection box will be disabled because the one has to be selected if it
exists. The steps of the selected rule will be shown in a help box to
finally show what the selected crush rule set does and the applied size
rules will be shown under the size input field.

Show the steps of the selected crush rule set the package
angular-animate is required, which is required by angular-bootstrap.

Amended all involved end-to-end tests.

Commit 3:
Calculate the best pg size for a pool (OP-2377)

Finally each pg number will match a number that is a power of two. The
min size of PGs that can be set is set to 16. It's only a simple
calculation, with the real values provided by the crushmap. The
calculation scheme is the following:

For replicated pools:

100 * $numOfOsds
------------------
    $size

For erasure coded pools:

100 * $numOfOsds
------------------
    $k + $m

The result will be rounded and converted to the next value that is a
power of two.

Signed-off-by: Stephan Müller <smueller@suse.com>