Packaging: Removes Nagios and PNP4Nagios from the RPM package (OP-2267)
As we've replaced the built-in monitoring/graphing based on
Nagios/Icinga and PNP4Nagios with an external solution based on
Prometheus and Grafana, there is no point in installing and configuring
the old monitoring infrastructure anymore.

Remove all related files and dependencies from the SUSE RPM spec file
and remove the script that is called during `oaconfig install` to
configure and enable Nagios.

The EL7 RPM spec file and Debian packaging have not been updated; they
will have to be overhauled completely anyway (see OP-2125 and OP-1950)

Signed-off-by: Lenz Grimmer <lenz@openattic.org>