Fix element not clickable in task-queue suite (OP-2262)
A strange problem occurred in the task queue suite, the tab element
was somehow not clickable for protractor, even so it was there and no
element was visible in front of it. It's possible that this happens
through the newly added angular's animation library or because it was
opened with the Grafana dashboard in background or because the
element was not updated in time. I came to the conclusion that there
is some truth in all assumptions. Now all task queue suites will be
opened on top of the OSDs table, the refresh button is used
instead of clicking through the UI tabs because of the animation
delay and it's fetching the element inside the test to prevent it
from being not up to date when it is needed.

The other suites had to be adjusted to reflect the changes done in the
common library.

Signed-off-by: Stephan Müller <smueller@suse.com>