Rename openattic RPM spec files (OP-1950)
In preparation of the consilidation of the SUSE and EL7 RPM spec files included in the git repo, rename the SUSE RPM spec file `openattic.spec.SUSE` to `openattic.spec`, as it's the current reference for the migration. Keep the EL7 RPM spec file around as `openattic.spec.EL7` for reference until the consolidation has been performed.

Signed-off-by: Lenz Grimmer <lenz@openattic.org>