Add user ID and keyring file info to RBD command (OP-2553)
We need to add at least the user ID and optional the keyring file
information to the RBD command. If the user ID is not given the RBD
command will use the default keyring file (ceph.client.admin.keyring)
which is not the correct one in environments set up by DeepSea for
example.

By providing the keyring file name as well as the user ID we're
skipping the algorithm to detected the related keyring for the given
user ID.

Signed-off-by: Tatjana Dehler <tdehler@suse.com>