Packaging: replace the `Version:` tag in the RPM spec file (OP-2256)
When building the source tar archive with `make_dist.py`, the script will now replace the `Version:` tag with the full version number including the timestamp for snapshot builds.

This ensures that a snapshot RPM build can be performed on the fly from the tarball, without having to adapt the RPM version manually.

To achieve this, the functionality provided by the `_get_build_basename()` method was moved from the `build()` method into the `_create_source_tarball()` method, which was the only method consuming the resulting `build_basename` anyway.

The arguments passed to `_create_source_tarball()` were adjusted accordingly.

Signed-off-by: Lenz Grimmer <lenz@openattic.org>