Unstuck RBD and pool creation tests (OP-2574)
The problem was that the unsaved changes dialog popped up in the RBD and
pool creations if the creation had failed because the tests wanted to
move on, but they couldn't. Now the unsaved changes dialog will be
handled in such a case and this will provide a clean up for the already
created pool/RBD because the other tests of those suites will not fail
anymore.

Signed-off-by: Stephan Müller <smueller@suse.com>