systemd: /usr/lib/systemd/system/openattic-systemd.service:9 Not an absolute path, ignoring: ${SYSD_PIDFILE} (OP-2570)
Remove the `PIDFile` directive out of the systemd service as it hasn't
been used.  The `EnvironmentFile` and `Environment` directive provide
environment variables for the `ExecStart` directive, but not for the
`PIDFile`.  Also removes the init.d script as this one is either not
used and very likely won't be used in the future due to the improved
integration of systemd in recent Linux distributions.

Signed-off-by: Patrick Nawracay <pnawracay@suse.com>