Upgrade Protractor (OP-2618)
With the upgrade of Protractor we also get a more recent version of webdriver and jasmine.

Add Protractor as a npm dev dependency.
With this  the correct versions of protractor, jasmine and webdriver will be installed when we run "npm install" on webui folder.

Add two new npm scripts: webdriver and protractor.
This allow us execute webdriver and protractor locally by running "npm run webdriver" and "npm run protractor".

Edit some tests so they can run on newer versions of protractor.

Signed-off-by: Tiago Melo <tmelo@suse.com>