Improve error handling of MON-widget (OP-2658)
Problem was that if no time check health for a MON existed, the
widget assumed that the MON was in bad health. Now if no health exists
the overall health status of the cluster is used. To display more
details on each MON the new detail error messages will be searched if
there is a mention of the MON name or address.

Signed-off-by: Stephan Müller <smueller@suse.com>