Fix task queue waiting errors (OP-2629)
Errors appeared that the reload button was not clickable because another
element would have received the click. This other element was surely a
notification that appeared on top of the refresh button. Now protractor
will click the position even if it will hit the notification.

Through the webpack change, the local test environment is to fast for
protractor that's why I needed to add sleeps.

Signed-off-by: Stephan Müller <smueller@suse.com>