Integrate the oA module_status into the Ceph app (OP-2687)
The oA module_status check will check if all preconditions of the oA Ceph
module are reached before opening a page that is related to the module.
Because the oA Ceph status check has NOT been added to the settings page,
the settings page will be accessible either with or without valid keyring
file.

Signed-off-by: Tatjana Dehler <tdehler@suse.com>