Add erasure coded profile e2e test suite (OP-2761)
Now the profile creation and deletion will be tested, too. Through
this missing suite we had a bug causing a JS error when you typed in a
profile name.

The error was appearing through a structure change in the
erasure coded profile component. The structure change was not applied to
the unique name directive which is used to determine if the name is
already taken.

Signed-off-by: Stephan Müller <smueller@suse.com>