Documentation: Enhanced installation instructions on Leap
Updated quick start guide to reflect the fact that Leap 42.3 ships with an incompatible version of Salt. Added instructions on how to downgrade and lock the version. Added pointer to the DeepSea Wiki and fixed some zypper repo URLs that used https (adding them failed with an obscure curl error).

Signed-off-by: Lenz Grimmer <lenz@openattic.org>