WebUI/QA: Clear input fields manually (OP-3187)
The problem was that we used protractors "clear" method to clear input
fields, this worked pretty well on most input fields but not with the
RBD object size field, if you clear it, it will be replaced with the
lowest object size value ('4.00 KiB'), if you than try to set the new
content you got something like ('4.00 KiB4.00 MiB'). This however was
nearly instantly converted to '4.00 KiB' again.

To get it working on this input field to the e2e test will first use
protractors clear as usually than send 'ctrl + a' to select any text
in the input field, than send text wanted,
which will replace the selected content.

Signed-off-by: Stephan Müller <smueller@suse.com>