WebUI/QA: Add browser sleep to iSCSI suite (OP-3190)
The problem was that a random failure occurs running iSCSI suites on
slower systems. A error will appear that checks if an updating text is
equal to the final text result. To prevent this, the browser should
sleep in order to have rendered the panel after it's sleep.

The suite also now uses "helpers.changeInput" function, instead of
calling a combination of protractor functions directly (which could also be an issue).

Signed-off-by: Stephan Müller <smueller@suse.com>