WebUI/QA: Refactore iSCSI suite (OP-3191)
The panel checking seemed to be the problem which couldn't be reproduced
(again see OP-3190). Now the iSCSI suite is refactored and a unit test
suite was added to check that the panel title changes on input change.
The panel checking was removed from the e2e test and tests that didn't
need any separation were combined. Created local functions and variables
to make the suite easier to read and to remove redundancy.

Signed-off-by: Stephan Müller <smueller@suse.com>