Backend: Return UTC timestamp for RBD snapshots (OP-3192)
Now the RBD snapshot timestamp will be send out by the backend with a
timezone marking it as UTC time. This will than be transformed into
the correct local time for the frontend user.

Explanation:

The frontend did everything correct, trying to format the time into the
local time.

The backend used the method "datetime.isoformat()" which assumed a
time format that resembles the ISO 8601 standard, but Python hasn't
implemented the standard correctly for this method
(https://bugs.python.org/issue23332).

One problem is that Python used naive datetimes only, up to version 3.2
(https://bugs.python.org/issue9527),
which have no knowledge about any timezone. In version 3.6 it mostly uses
time zone aware datetimes.

The second problem is that Librados uses
"datetime.utcfromtimestamp($timestamp)"
method, which will even in Python 3.6 return a naive datetime.
There is a workaround for the call is using
"datetime.fromtimestamp($timestamp, timezone.utc)",
but this will only work with Python 3.2 or higher, as timezone hasn't
existed before.

The third kind of problem is that, as mentioned before, "datetime.isoformat"
is not correctly implement, even in Python 3.6 a time zone aware datetime
won't be append with the time zone suffix, only if you call
"datetime.astimezone().isoformat()".

Looking at the second problem, it's clear that the method will not
change soon, but even when it changes to be time zone aware, the simple
prefix "Z", to make it clear to the frontend that it's an UTC time,
won't be broken as the third problem will exist in python 3.x (no change
expected ATM).

Signed-off-by: Stephan Müller <smueller@suse.com>