Backend: Create pools with rule (OP-3200)
The problem was that the pools were first created without a rule, after
creation the pool rule was changed. This wasn't a problem until
a rule with the same name as the created pool was automatically created
if a erasure coded profile was submitted.
This cli convenience caused this bug, as it created rules with ec pool
names on every creation of an ec pool that didn't used the 'default'
ec profile.

Signed-off-by: Stephan Müller <smueller@suse.com>